+++ filesfolders/views.py
@@ -21,7 +21,7 @@ from .utils import build_public_url
 # Projectroles dependency
 from projectroles.models import Project, ProjectSetting
 from projectroles.plugins import get_backend_api
-from projectroles.utils import get_project_setting
+from projectroles.project_settings import get_project_setting
 from projectroles.views import LoggedInPermissionMixin, \
     ProjectContextMixin